+++ scanpy/plotting/_anndata.py
@@ -2127,241 +2127,6 @@ def _prepare_dataframe(
     return categories, obs_tidy
 
 
-def _plot_gene_groups_brackets(
-    gene_groups_ax: Axes,
-    group_positions: Iterable[Tuple[int, int]],
-    group_labels: Sequence[str],
-    left_adjustment: float = -0.3,
-    right_adjustment: float = 0.3,
-    rotation: Optional[float] = None,
-    orientation: Literal['top', 'right'] = 'top',
-):
-    """\
-    Draws brackets that represent groups of genes on the give axis.
-    For best results, this axis is located on top of an image whose
-    x axis contains gene names.
-
-    The gene_groups_ax should share the x axis with the main ax.
-
-    Eg: gene_groups_ax = fig.add_subplot(axs[0, 0], sharex=dot_ax)
-
-    This function is used by dotplot, heatmap etc.
-
-    Parameters
-    ----------
-    gene_groups_ax
-        In this axis the gene marks are drawn
-    group_positions
-        Each item in the list, should contain the start and end position that the
-        bracket should cover.
-        Eg. [(0, 4), (5, 8)] means that there are two brackets, one for the var_names (eg genes)
-        in positions 0-4 and other for positions 5-8
-    group_labels
-        List of group labels
-    left_adjustment
-        adjustment to plot the bracket start slightly before or after the first gene position.
-        If the value is negative the start is moved before.
-    right_adjustment
-        adjustment to plot the bracket end slightly before or after the last gene position
-        If the value is negative the start is moved before.
-    rotation
-        rotation degrees for the labels. If not given, small labels (<4 characters) are not
-        rotated, otherwise, they are rotated 90 degrees
-    orientation
-        location of the brackets. Either `top` or `right`
-    Returns
-    -------
-    None
-    """
-    import matplotlib.patches as patches
-    from matplotlib.path import Path
-
-    # get the 'brackets' coordinates as lists of start and end positions
-
-    left = [x[0] + left_adjustment for x in group_positions]
-    right = [x[1] + right_adjustment for x in group_positions]
-
-    # verts and codes are used by PathPatch to make the brackets
-    verts = []
-    codes = []
-    if orientation == 'top':
-        # rotate labels if any of them is longer than 4 characters
-        if rotation is None and group_labels:
-            if max([len(x) for x in group_labels]) > 4:
-                rotation = 90
-            else:
-                rotation = 0
-        for idx in range(len(left)):
-            verts.append((left[idx], 0))  # lower-left
-            verts.append((left[idx], 0.6))  # upper-left
-            verts.append((right[idx], 0.6))  # upper-right
-            verts.append((right[idx], 0))  # lower-right
-
-            codes.append(Path.MOVETO)
-            codes.append(Path.LINETO)
-            codes.append(Path.LINETO)
-            codes.append(Path.LINETO)
-
-            try:
-                group_x_center = left[idx] + float(right[idx] - left[idx]) / 2
-                gene_groups_ax.text(
-                    group_x_center,
-                    1.1,
-                    group_labels[idx],
-                    ha='center',
-                    va='bottom',
-                    rotation=rotation,
-                )
-            except:
-                pass
-    else:
-        top = left
-        bottom = right
-        for idx in range(len(top)):
-            verts.append((0, top[idx]))  # upper-left
-            verts.append((0.15, top[idx]))  # upper-right
-            verts.append((0.15, bottom[idx]))  # lower-right
-            verts.append((0, bottom[idx]))  # lower-left
-
-            codes.append(Path.MOVETO)
-            codes.append(Path.LINETO)
-            codes.append(Path.LINETO)
-            codes.append(Path.LINETO)
-
-            try:
-                diff = bottom[idx] - top[idx]
-                group_y_center = top[idx] + float(diff) / 2
-                if diff * 2 < len(group_labels[idx]):
-                    # cut label to fit available space
-                    group_labels[idx] = group_labels[idx][: int(diff * 2)] + "."
-                gene_groups_ax.text(
-                    0.6,
-                    group_y_center,
-                    group_labels[idx],
-                    ha='right',
-                    va='center',
-                    rotation=270,
-                    fontsize='small',
-                )
-            except Exception as e:
-                print('problems {}'.format(e))
-                pass
-
-    path = Path(verts, codes)
-
-    patch = patches.PathPatch(path, facecolor='none', lw=1.5)
-
-    gene_groups_ax.add_patch(patch)
-    gene_groups_ax.grid(False)
-    gene_groups_ax.axis('off')
-    # remove y ticks
-    gene_groups_ax.tick_params(axis='y', left=False, labelleft=False)
-    # remove x ticks and labels
-    gene_groups_ax.tick_params(
-        axis='x', bottom=False, labelbottom=False, labeltop=False
-    )
-
-
-def _reorder_categories_after_dendrogram(
-    adata: AnnData,
-    groupby,
-    dendrogram,
-    var_names=None,
-    var_group_labels=None,
-    var_group_positions=None,
-):
-    """\
-    Function used by plotting functions that need to reorder the the groupby
-    observations based on the dendrogram results.
-
-    The function checks if a dendrogram has already been precomputed.
-    If not, `sc.tl.dendrogram` is run with default parameters.
-
-    The results found in `.uns[dendrogram_key]` are used to reorder
-    `var_group_labels` and `var_group_positions`.
-
-
-    Returns
-    -------
-    dictionary with keys:
-    'categories_idx_ordered', 'var_group_names_idx_ordered',
-    'var_group_labels', and 'var_group_positions'
-    """
-
-    key = _get_dendrogram_key(adata, dendrogram, groupby)
-
-    dendro_info = adata.uns[key]
-    if groupby != dendro_info['groupby']:
-        raise ValueError(
-            "Incompatible observations. The precomputed dendrogram contains "
-            f"information for the observation: '{groupby}' while the plot is "
-            f"made for the observation: '{dendro_info['groupby']}. "
-            "Please run `sc.tl.dendrogram` using the right observation.'"
-        )
-
-    categories = adata.obs[dendro_info['groupby']].cat.categories
-
-    # order of groupby categories
-    categories_idx_ordered = dendro_info['categories_idx_ordered']
-    categories_ordered = dendro_info['categories_ordered']
-
-    if len(categories) != len(categories_idx_ordered):
-        raise ValueError(
-            "Incompatible observations. Dendrogram data has "
-            f"{len(categories_idx_ordered)} categories but current groupby "
-            f"observation {groupby!r} contains {len(categories)} categories. "
-            "Most likely the underlying groupby observation changed after the "
-            "initial computation of `sc.tl.dendrogram`. "
-            "Please run `sc.tl.dendrogram` again.'"
-        )
-
-    # reorder var_groups (if any)
-    if var_names is not None:
-        var_names_idx_ordered = list(range(len(var_names)))
-
-    if var_group_positions:
-        if set(var_group_labels) == set(categories):
-            positions_ordered = []
-            labels_ordered = []
-            position_start = 0
-            var_names_idx_ordered = []
-            for cat_name in categories_ordered:
-                idx = var_group_labels.index(cat_name)
-                position = var_group_positions[idx]
-                _var_names = var_names[position[0] : position[1] + 1]
-                var_names_idx_ordered.extend(range(position[0], position[1] + 1))
-                positions_ordered.append(
-                    (position_start, position_start + len(_var_names) - 1)
-                )
-                position_start += len(_var_names)
-                labels_ordered.append(var_group_labels[idx])
-            var_group_labels = labels_ordered
-            var_group_positions = positions_ordered
-        else:
-            logg.warning(
-                "Groups are not reordered because the `groupby` categories "
-                "and the `var_group_labels` are different.\n"
-                f"categories: {_format_first_three_categories(categories)}\n"
-                f"var_group_labels: {_format_first_three_categories(var_group_labels)}"
-            )
-    else:
-        var_names_idx_ordered = None
-
-    if var_names_idx_ordered is not None:
-        var_names_ordered = [var_names[x] for x in var_names_idx_ordered]
-    else:
-        var_names_ordered = None
-
-    return dict(
-        categories_idx_ordered=categories_idx_ordered,
-        categories_ordered=dendro_info['categories_ordered'],
-        var_names_idx_ordered=var_names_idx_ordered,
-        var_names_ordered=var_names_ordered,
-        var_group_labels=var_group_labels,
-        var_group_positions=var_group_positions
-    )
-
-
 def _format_first_three_categories(categories):
     categories = list(categories)
     if len(categories) > 3:
